+++ hash_extender_sha1.c
@@ -12,7 +12,7 @@
 #define SHA1_BLOCK 64
 
 /* Note: this only supports data with a 4-byte size (4.2 billion bits). */
-uint8_t *sha1_append_data(uint8_t *data, uint64_t data_length, size_t secret_length, uint8_t *append, size_t append_length, size_t *new_length)
+uint8_t *sha1_append_data(uint8_t *data, uint64_t data_length, uint64_t secret_length, uint8_t *append, uint64_t append_length, uint64_t *new_length)
 {
   /* Allocate memory for the new buffer (enough room for buffer + a full block + the data) */
   uint8_t *result = (uint8_t*) malloc(1000 + data_length + append_length + SHA1_BLOCK); /* (This can overflow if we're ever using this in a security-sensitive context) */
@@ -46,7 +46,7 @@ uint8_t *sha1_append_data(uint8_t *data, uint64_t data_length, size_t secret_len
   return result;
 }
 
-void sha1_gen_signature(uint8_t *secret, uint64_t secret_length, uint8_t *data, size_t data_length, uint8_t signature[SHA_DIGEST_LENGTH])
+void sha1_gen_signature(uint8_t *secret, uint64_t secret_length, uint8_t *data, uint64_t data_length, uint8_t signature[SHA_DIGEST_LENGTH])
 {
   SHA_CTX c;
   SHA1_Init(&c);
@@ -55,7 +55,7 @@ void sha1_gen_signature(uint8_t *secret, uint64_t secret_length, uint8_t *data, 
   SHA1_Final(signature, &c);
 }
 
-void sha1_gen_signature_evil(uint64_t secret_length, size_t data_length, uint8_t original_signature[SHA_DIGEST_LENGTH], uint8_t *append, size_t append_length, uint8_t new_signature[SHA_DIGEST_LENGTH])
+void sha1_gen_signature_evil(uint64_t secret_length, uint64_t data_length, uint8_t original_signature[SHA_DIGEST_LENGTH], uint8_t *append, uint64_t append_length, uint8_t new_signature[SHA_DIGEST_LENGTH])
 {
   SHA_CTX c;
   uint64_t original_data_length;
@@ -84,7 +84,7 @@ void sha1_gen_signature_evil(uint64_t secret_length, size_t data_length, uint8_t
   SHA1_Final(new_signature, &c);
 }
 
-static int sha1_test_validate(uint8_t *secret, uint64_t secret_length, uint8_t *data, size_t data_length, uint8_t *signature)
+static int sha1_test_validate(uint8_t *secret, uint64_t secret_length, uint8_t *data, uint64_t data_length, uint8_t *signature)
 {
   unsigned char result[SHA_DIGEST_LENGTH];
